feat(metasrv): metasrv unittest logs tracing event with customized formatter
Move the logging config from common/tracing to metasrv/:
`metasrv` sometimes needs full span tracing to debug. It's a little
different from the logging requirement for query.

- Revert meta ut logging config removed from #6845
