+++ metasrv/tests/it/tests/service.rs
@@ -208,7 +208,7 @@ macro_rules! init_meta_ut {
         let t = tempfile::tempdir().expect("create temp dir to sled db");
         common_meta_sled_store::init_temp_sled_db(t);
 
-        common_tracing::init_logging("meta_unittests", &common_tracing::Config::new_testing());
+        $crate::tests::logging::init_meta_ut_tracing();
 
         let name = common_tracing::func_name!();
         let span = tracing::debug_span!("ut", "{}", name.split("::").last().unwrap());
